Add catcher same mesage

+++ bot/utils/fsm/window_utility.py
@@ -1,11 +1,15 @@
+import logging
 from typing import Any
 
 from aiogram import Bot
+from aiogram.exceptions import TelegramBadRequest
 from aiogram.fsm.context import FSMContext
-from aiogram.types import InputMediaPhoto
+from aiogram.types import InputMediaPhoto, CallbackQuery
 
 from utils.fsm.fsm_utility import MessageType
 from utils.update import get_chat_id
+
+logger = logging.getLogger(__name__)
 
 
 async def remove_window_message(ctx: Any, bot: Bot, state: FSMContext):
@@ -14,8 +18,9 @@ async def remove_window_message(ctx: Any, bot: Bot, state: FSMContext):
     message_key = next((k for k, v in data.items() if k.endswith(MessageType.Window)), None)
     if message_key:
         message_id = data[message_key]
-        await bot.delete_message(get_chat_id(ctx), message_id=message_id)
-        await state.update_data({message_key: None})
+        if message_id:
+            await bot.delete_message(get_chat_id(ctx), message_id=message_id)
+            await state.update_data({message_key: None})
 
 
 async def window_info(ctx: Any, bot: Bot, state: FSMContext, *, prefix: str, **func_kwargs):
@@ -36,7 +41,17 @@ async def window_info(ctx: Any, bot: Bot, state: FSMContext, *, prefix: str, **f
     bot_kwargs.pop(message_key, None)
 
     if main_message_id:
-        await edit_main_message(chat_id, main_message_id, bot, **bot_kwargs)
+        try:
+            await edit_main_message(chat_id, main_message_id, bot, **bot_kwargs)
+        except Exception as exception:
+            if isinstance(exception, TelegramBadRequest):
+                if "Bad Request: message is not modified:" in exception.message:
+                    logger.warning("Bad Request: message is not modified:")
+                    if isinstance(ctx, CallbackQuery):
+                        await ctx.answer()
+                    return
+
+            raise
     else:
         await send_main_message(ctx, bot, state, prefix=prefix, **bot_kwargs)
